Move SummaryTimeline into /Edit/Components/SummaryTimeline
